fix(security): polish/sigma-security-p1p2 — Omega 12 patches, MCP security hardening
Squash-merges polish/sigma-security-p1p2 (121 commits) onto main.
Conflict resolution: 43 files resolved (HEAD for src, keep-both for tests).

Security patches applied:
- MCP coherence_server: skill validation + fence-escape sanitization
- MCP hookify_server: SurrealQL injection fix
- MCP marimo report: RCE elimination + drop shell=True
- auth.py: formatting cleanup
- DynamicConcurrencyGate singleton reset in conftest.py
- 3 xfail markers for pending SurrealDB/vault integration in test_hookify.py

Also includes Wave 2E/2F type and security annotations across 100+ files.

Entire-Checkpoint: eaa36ebf1a9e

+++ src/cohezion/swarm/anomaly_detector.py
@@ -145,7 +145,7 @@ class ModelCostHistory:
         if denominator == 0:
             return 0.0
 
-        return numerator / denominator  # Cost per second
+        return float(numerator / denominator)  # Cost per second
 
 
 class AnomalyDetector:
